requirement.txt
asgiref==3.4.1
Django==3.2.9
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
djangorestframework
drf_yasg